2024 06 01

+++ app/src/main/java/es/jvbabi/vplanplus/feature/main_homework/shared/data/repository/HomeworkRepositoryImpl.kt
@@ -14,7 +14,7 @@ import es.jvbabi.vplanplus.data.source.database.converter.ZonedDateTimeConverter
 import es.jvbabi.vplanplus.data.source.database.dao.HomeworkDao
 import es.jvbabi.vplanplus.data.source.database.dao.PreferredHomeworkNotificationTimeDao
 import es.jvbabi.vplanplus.domain.model.Classes
-import es.jvbabi.vplanplus.domain.model.VppId
+import es.jvbabi.vplanplus.domain.model.Profile
 import es.jvbabi.vplanplus.domain.repository.ClassRepository
 import es.jvbabi.vplanplus.domain.repository.DefaultLessonRepository
 import es.jvbabi.vplanplus.domain.repository.KeyValueRepository
@@ -28,6 +28,7 @@ import es.jvbabi.vplanplus.domain.repository.VppIdRepository
 import es.jvbabi.vplanplus.feature.main_homework.shared.data.model.DbPreferredNotificationTime
 import es.jvbabi.vplanplus.feature.main_homework.shared.domain.model.Homework
 import es.jvbabi.vplanplus.feature.main_homework.shared.domain.model.HomeworkTask
+import es.jvbabi.vplanplus.feature.main_homework.shared.domain.repository.DeleteTask
 import es.jvbabi.vplanplus.feature.main_homework.shared.domain.repository.HomeworkModificationResult
 import es.jvbabi.vplanplus.feature.main_homework.shared.domain.repository.HomeworkRepository
 import es.jvbabi.vplanplus.feature.main_homework.shared.domain.repository.NewTaskRecord
@@ -70,7 +71,7 @@ class HomeworkRepositoryImpl(
         profileRepository
             .getProfiles()
             .first()
-            .filter { it.type == ProfileType.STUDENT }
+            .filter { it.type == ProfileType.STUDENT && it.isHomeworkEnabled }
             .forEach { profile ->
                 val `class` = classRepository.getClassById(profile.referenceId) ?: return@forEach
                 val school = `class`.school
@@ -151,7 +152,7 @@ class HomeworkRepositoryImpl(
 
                     insertHomework(
                         id = id,
-                        createdBy = createdBy,
+                        profile = profile,
                         shareWithClass = responseHomework.shareWithClass,
                         until = until,
                         `class` = `class`,
@@ -159,7 +160,7 @@ class HomeworkRepositoryImpl(
                         createdAt = ZonedDateTimeConverter().timestampToZonedDateTime(
                             responseHomework.createdAt
                         ),
-                        allowCloudUpdate = false,
+                        storeInCloud = false,
                         tasks = replacementTasks,
                         isHidden = (existingRecord?.isHidden ?: (isNewHomework && until.isBefore(ZonedDateTime.now())) && createdBy?.id != profile.vppId?.id),
                     )
@@ -250,26 +251,27 @@ class HomeworkRepositoryImpl(
 
     override suspend fun insertHomework(
         id: Long?,
-        createdBy: VppId?,
-        createdAt: ZonedDateTime,
+        profile: Profile,
         `class`: Classes,
         defaultLessonVpId: Long?,
+        storeInCloud: Boolean,
         shareWithClass: Boolean,
         until: ZonedDateTime,
         tasks: List<NewTaskRecord>,
-        allowCloudUpdate: Boolean,
-        isHidden: Boolean
+        isHidden: Boolean,
+        createdAt: ZonedDateTime
     ): HomeworkModificationResult {
-        if (!allowCloudUpdate || createdBy == null) {
+        if (!storeInCloud || profile.vppId == null) {
             val dbHomework = DbHomework(
                 id = id ?: (findLocalId() - 1),
                 classes = `class`.classId,
                 createdAt = createdAt,
                 until = until,
                 defaultLessonVpId = defaultLessonVpId,
-                createdBy = createdBy?.id,
+                createdBy = profile.vppId?.id,
                 hidden = isHidden,
-                isPublic = shareWithClass
+                isPublic = shareWithClass,
+                profileId = profile.id
             )
             homeworkDao.insert(dbHomework)
             tasks.forEach { newTask ->
@@ -284,8 +286,7 @@ class HomeworkRepositoryImpl(
             return HomeworkModificationResult.SUCCESS_OFFLINE
         }
 
-        val vppIdToken =
-            vppIdRepository.getVppIdToken(createdBy) ?: return HomeworkModificationResult.FAILED
+        val vppIdToken = vppIdRepository.getVppIdToken(profile.vppId) ?: return HomeworkModificationResult.FAILED
         vppIdNetworkRepository.authentication = BearerAuthentication(vppIdToken)
         val result = vppIdNetworkRepository.doRequest(
             path = "/api/$API_VERSION/user/me/homework",
@@ -308,9 +309,10 @@ class HomeworkRepositoryImpl(
             createdAt = createdAt,
             until = until,
             defaultLessonVpId = defaultLessonVpId,
-            createdBy = createdBy.id,
+            createdBy = profile.vppId.id,
             isPublic = shareWithClass,
-            hidden = isHidden
+            hidden = isHidden,
+            profileId = profile.id
         )
         homeworkDao.insert(dbHomework)
         response.tasks.forEach {
@@ -325,32 +327,44 @@ class HomeworkRepositoryImpl(
         return HomeworkModificationResult.SUCCESS_ONLINE_AND_OFFLINE
     }
 
-    override suspend fun deleteOrHideHomework(
+    override suspend fun removeOrHideHomework(
         homework: Homework,
-        onlyHide: Boolean
+        task: DeleteTask
     ): HomeworkModificationResult {
-        if (homework.id < 0) {
-            homeworkDao.deleteHomework(homework.id)
-            return HomeworkModificationResult.SUCCESS_OFFLINE
-        }
 
-        val vppId = vppIdRepository
-            .getVppIds().first()
-            .firstOrNull { it.isActive() && it.id == homework.createdBy?.id }
-            ?: return HomeworkModificationResult.FAILED
+        when (task) {
+            DeleteTask.HIDE -> {
+                homeworkDao.changeHidden(homework.id, true)
+                return HomeworkModificationResult.SUCCESS_OFFLINE
+            }
+            DeleteTask.DELETE -> {
+                if (homework.id < 0) {
+                    homeworkDao.deleteHomework(homework.id)
+                    return HomeworkModificationResult.SUCCESS_OFFLINE
+                }
+                val vppId = vppIdRepository
+                    .getVppIds().first()
+                    .firstOrNull { it.isActive() && it.id == homework.createdBy?.id }
+                    ?: return HomeworkModificationResult.FAILED
 
-        val vppIdToken =
-            vppIdRepository.getVppIdToken(vppId) ?: return HomeworkModificationResult.FAILED
-        vppIdNetworkRepository.authentication = BearerAuthentication(vppIdToken)
-        val result = vppIdNetworkRepository.doRequest(
-            path = "/api/$API_VERSION/user/me/homework/${homework.id}",
-            requestMethod = HttpMethod.Delete
-        )
-        return if (result.response == HttpStatusCode.OK) {
-            homeworkDao.deleteHomework(homework.id)
-            HomeworkModificationResult.SUCCESS_OFFLINE
-        } else {
-            HomeworkModificationResult.FAILED
+                val vppIdToken =
+                    vppIdRepository.getVppIdToken(vppId) ?: return HomeworkModificationResult.FAILED
+                vppIdNetworkRepository.authentication = BearerAuthentication(vppIdToken)
+                val result = vppIdNetworkRepository.doRequest(
+                    path = "/api/$API_VERSION/user/me/homework/${homework.id}",
+                    requestMethod = HttpMethod.Delete
+                )
+                return if (result.response == HttpStatusCode.OK) {
+                    homeworkDao.deleteHomework(homework.id)
+                    HomeworkModificationResult.SUCCESS_OFFLINE
+                } else {
+                    HomeworkModificationResult.FAILED
+                }
+            }
+            DeleteTask.FORCE_DELETE_LOCALLY -> {
+                homeworkDao.deleteHomework(homework.id)
+                return HomeworkModificationResult.SUCCESS_OFFLINE
+            }
         }
     }
 
@@ -534,11 +548,6 @@ class HomeworkRepositoryImpl(
         } else {
             HomeworkModificationResult.FAILED
         }
-    }
-
-    override suspend fun changeVisibility(homework: Homework): HomeworkModificationResult {
-        homeworkDao.changeHidden(homework.id, !homework.isHidden)
-        return HomeworkModificationResult.SUCCESS_OFFLINE
     }
 
     override suspend fun updateDueDate(
